requirements.txt
flask
pyautogui
mss
goto-statement
numpy
scipy
Pillow
pypiwin32==223
pywin32==224
scikit-image
opencv-python
